+++ .gitignore
@@ -19,7 +19,6 @@ examples/models_training/my_model/*
 examples/notebooks/models_training/model_generation.ipynb
 examples/notebooks/models_training/generated/*
 examples/scripts/my_models*
-examples/scripts/reproducibility
 examples/notebooks/models_training/my_model/
 examples/showcases copy/*
 examples/notebooks/models_training/generate.py
@@ -27,7 +26,8 @@ dummy_output_dir/*
 examples/notebooks/models_training/my_model/*
 examples/notebooks/test.ipynb
 examples/notebooks/dummy_output_dir/*
-examples/scripts/reproducibility/*
+examples/scripts/reproducibility/my_models
+examples/scripts/reproducibility/data
 wandb
 examples/notebooks/dummy_output_dir/
 examples/notebooks/models_training/dummy_output_dir/
@@ -50,6 +50,7 @@ examples/scripts/plots/*
 examples/notebooks/my_model_with_custom_archi/
 examples/notebooks/my_model/
 examples/net.py
+examples/scripts/configs/*
