+++ examples/scripts/reproducibility/configs/binary_mnist/hvae/hvae_config.json
@@ -1,0 +1,10 @@
+{
+    "name": "HVAEConfig",
+    "latent_dim": 64,
+    "reconstruction_loss": "bce",
+    "n_lf": 5,
+    "eps_lf": 0.001,
+    "beta_zero": 1,
+    "learn_eps_lf": false,
+    "learn_beta_zero": false
+}
