+++ examples/scripts/reproducibility/configs/binary_mnist/iwae/base_training_config.json
@@ -1,0 +1,10 @@
+{
+  "name": "BaseTrainerConfig",
+  "output_dir": "reproducibility/binary_mnist",
+  "batch_size": 20,
+  "num_epochs": 3280,
+  "learning_rate": 1e-3,
+  "steps_saving": null,
+  "steps_predict": null,
+  "no_cuda": false
+}
