+++ examples/scripts/reproducibility/configs/binary_mnist/vae/base_training_config.json
@@ -1,0 +1,10 @@
+{
+  "name": "BaseTrainerConfig",
+  "output_dir": "reproducibility/binary_mnist",
+  "batch_size": 100,
+  "num_epochs": 500,
+  "learning_rate": 1e-4,
+  "steps_saving": null,
+  "steps_predict": null,
+  "no_cuda": false
+}
