+++ examples/scripts/reproducibility/configs/binary_mnist/vae/vae_config.json
@@ -1,0 +1,5 @@
+{
+    "name": "VAEConfig",
+    "latent_dim": 40,
+    "reconstruction_loss": "bce"
+}
