+++ examples/scripts/reproducibility/configs/binary_mnist/vamp/vamp_config.json
@@ -1,0 +1,6 @@
+{
+    "name": "VAMPConfig",
+    "latent_dim": 40,
+    "reconstruction_loss": "bce",
+    "number_components": 500
+}
