+++ examples/scripts/reproducibility/configs/celeba/rae_gp/base_training_config.json
@@ -1,0 +1,10 @@
+{
+  "name": "BaseTrainerConfig",
+  "output_dir": "my_models_on_celeba",
+  "batch_size": 100,
+  "num_epochs": 55,
+  "learning_rate": 0.001,
+  "steps_saving": null,
+  "steps_predict": 100,
+  "no_cuda": false
+}
