+++ examples/scripts/reproducibility/configs/celeba/rae_gp/rae_gp_config.json
@@ -1,0 +1,6 @@
+{
+    "name": "RAE_L2_Config",
+    "latent_dim": 64,
+    "embedding_weight": 1e-4,
+    "reg_weight": 1e-7
+  }
