+++ examples/scripts/reproducibility/configs/celeba/wae/base_training_config.json
@@ -1,0 +1,10 @@
+{
+  "name": "BaseTrainerConfig",
+  "output_dir": "my_models_on_celeba",
+  "batch_size": 100,
+  "num_epochs": 100,
+  "learning_rate": 0.0001,
+  "steps_saving": 3,
+  "steps_predict": 100,
+  "no_cuda": false
+}
