+++ examples/scripts/reproducibility/configs/dsprites/beta_tc_vae/base_training_config.json
@@ -1,0 +1,10 @@
+{
+  "name": "BaseTrainerConfig",
+  "output_dir": "reproducibility/dsprites",
+  "batch_size": 1000,
+  "num_epochs": 50,
+  "learning_rate": 1e-3,
+  "steps_saving": null,
+  "steps_predict": null,
+  "no_cuda": false
+}
