+++ examples/scripts/reproducibility/configs/dsprites/beta_tc_vae/beta_tc_vae_config.json
@@ -1,0 +1,9 @@
+{
+    "name": "BetaTCVAEConfig",
+    "latent_dim": 10,
+    "reconstruction_loss": "bce",
+    "beta": 6,
+    "gamma": 1,
+    "alpha": 1,
+    "use_mss": false
+}
