+++ examples/scripts/reproducibility/configs/mnist/svae/base_training_config.json
@@ -1,0 +1,10 @@
+{
+  "name": "BaseTrainerConfig",
+  "output_dir": "reproducibility/binary_mnist",
+  "batch_size": 64,
+  "num_epochs": 500,
+  "learning_rate": 1e-3,
+  "steps_saving": 50,
+  "steps_predict": null,
+  "no_cuda": false
+}
