+++ examples/scripts/reproducibility/configs/mnist/svae/svae_config.json
@@ -1,0 +1,5 @@
+{
+    "name": "SVAEConfig",
+    "latent_dim": 11,
+    "reconstruction_loss": "bce"
+}
