+++ examples/scripts/reproducibility/svae.py
@@ -1,0 +1,266 @@
+import argparse
+import logging
+import os
+from typing import List
+
+import numpy as np
+import torch
+
+from pythae.data.preprocessors import DataProcessor
+from pythae.models import SVAE, SVAEConfig
+from pythae.models import AutoModel
+from pythae.trainers import BaseTrainerConfig, BaseTrainer
+
+from pythae.models.nn import BaseEncoder, BaseDecoder
+import torch.nn as nn
+from pythae.models.base.base_utils import ModelOutput
+
+
+logger = logging.getLogger(__name__)
+console = logging.StreamHandler()
+logger.addHandler(console)
+logger.setLevel(logging.INFO)
+
+PATH = os.path.dirname(os.path.abspath(__file__))
+
+ap = argparse.ArgumentParser()
+
+device = 'cuda' if torch.cuda.is_available() else 'cpu'
+
+
+ap.add_argument(
+    "--model_config",
+    help="path to model config file (expected json file)",
+    default=None,
+)
+ap.add_argument(
+    "--training_config",
+    help="path to training config_file (expected json file)",
+    default=os.path.join(PATH, "configs/base_training_config.json"),
+)
+
+args = ap.parse_args()
+
+### Define paper encoder network
+class Encoder(BaseEncoder):
+    def __init__(self, args: dict):
+        BaseEncoder.__init__(self)
+        self.input_dim = args.input_dim
+        self.latent_dim = args.latent_dim
+
+        layers = nn.ModuleList()
+
+        layers.append(
+            nn.Sequential(
+                nn.Linear(np.prod(args.input_dim), 256),
+                nn.ReLU(),
+                nn.Linear(256, 128),
+                nn.ReLU()
+            )
+        )
+
+        self.layers = layers
+        self.depth = len(layers)
+
+        self.embedding = nn.Linear(128, self.latent_dim)
+        self.log_concentration = nn.Linear(128, 1)
+
+
+    def forward(self, x, output_layer_levels: List[int] = None):
+        output = ModelOutput()
+
+        max_depth = self.depth
+
+        if output_layer_levels is not None:
+
+            assert all(
+                self.depth >= levels > 0 or levels == -1
+                for levels in output_layer_levels
+            ), (
+                f"Cannot output layer deeper than depth ({self.depth}). Got ({output_layer_levels})."
+            )
+
+            if -1 in output_layer_levels:
+                max_depth = self.depth
+            else:
+                max_depth = max(output_layer_levels)
+
+        out = x.reshape(-1, np.prod(self.input_dim))
+
+        for i in range(max_depth):
+            out = self.layers[i](out)
+
+            if output_layer_levels is not None:
+                if i + 1 in output_layer_levels:
+                    output[f"embedding_layer_{i+1}"] = out
+            if i + 1 == self.depth:
+                output["embedding"] = self.embedding(out)
+                output["log_concentration"] = self.log_concentration(out)
+
+        return output
+
+### Define paper decoder network
+class Decoder(BaseDecoder):
+    def __init__(self, args: dict):
+        BaseDecoder.__init__(self)
+
+        self.input_dim = args.input_dim
+
+        # assert 0, np.prod(args.input_dim)
+
+        layers = nn.ModuleList()
+
+        layers.append(
+            nn.Sequential(
+                nn.Linear(args.latent_dim, 128),
+                nn.ReLU(),
+                nn.Linear(128, 256),
+                nn.ReLU()
+            )
+        )
+
+        layers.append(
+            nn.Sequential(
+                nn.Linear(256, np.prod(args.input_dim)),
+                nn.Sigmoid()
+            )
+        )
+
+        self.layers = layers
+        self.depth = len(layers)
+
+    def forward(self, z: torch.Tensor, output_layer_levels: List[int] = None):
+
+        output = ModelOutput()
+
+        max_depth = self.depth
+
+        if output_layer_levels is not None:
+
+            assert all(
+                self.depth >= levels > 0 or levels == -1
+                for levels in output_layer_levels
+            ), (
+                f"Cannot output layer deeper than depth ({self.depth}). "
+                f"Got ({output_layer_levels})."
+            )
+
+            if -1 in output_layer_levels:
+                max_depth = self.depth
+            else:
+                max_depth = max(output_layer_levels)
+
+        out = z
+
+        for i in range(max_depth):
+            out = self.layers[i](out)
+
+            if output_layer_levels is not None:
+                if i + 1 in output_layer_levels:
+                    output[f"reconstruction_layer_{i+1}"] = out
+            if i + 1 == self.depth:
+                output["reconstruction"] = out.reshape((z.shape[0],) + self.input_dim)
+
+        return output
+
+
+
+def main(args):
+
+    ### Load data
+    train_data = torch.tensor(
+        np.load(os.path.join(PATH, f"data/mnist", "train_data.npz"))[
+            "data"
+        ]
+        / 255.0
+    )
+    eval_data = torch.tensor(
+        np.load(os.path.join(PATH, f"data/mnist", "eval_data.npz"))["data"]
+        / 255.0
+    )
+
+    train_data = torch.cat((train_data, eval_data))
+
+    test_data = (
+        np.load(os.path.join(PATH, f"data/mnist", "test_data.npz"))["data"]
+        / 255.0
+    )
+
+    data_input_dim = tuple(train_data.shape[1:])
+
+
+    if args.model_config is not None:
+        model_config = SVAEConfig.from_json_file(args.model_config)
+
+    else:
+        model_config = SVAEConfig()
+
+    model_config.input_dim = data_input_dim
+
+    model = SVAE(
+        model_config=model_config,
+        encoder=Encoder(model_config),
+        decoder=Decoder(model_config),
+    )
+
+    ### Set training config
+    training_config = BaseTrainerConfig.from_json_file(args.training_config)
+
+    ### Process data
+    data_processor = DataProcessor()
+    logger.info("Preprocessing train data...")
+    train_data = data_processor.process_data(torch.bernoulli(train_data))
+    train_dataset = data_processor.to_dataset(train_data)
+
+    logger.info("Preprocessing eval data...\n")
+    eval_data = data_processor.process_data(torch.bernoulli(eval_data))
+    eval_dataset = data_processor.to_dataset(eval_data)
+
+    ### Optimizer
+    optimizer = torch.optim.Adam(model.parameters(), lr=training_config.learning_rate)
+
+    ### Scheduler
+    scheduler = torch.optim.lr_scheduler.MultiStepLR(
+        optimizer, milestones=[10000000], gamma=10**(-1/3), verbose=True
+    )
+
+    seed = 123
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)
+
+    logger.info("Using Base Trainer\n")
+    trainer = BaseTrainer(
+        model=model,
+        train_dataset=train_dataset,
+        eval_dataset=None,#eval_dataset,
+        training_config=training_config,
+        optimizer=optimizer,
+        scheduler=scheduler,
+        callbacks=None,
+    )
+
+    ### Launch training
+    trainer.train()
+    
+    trained_model = AutoModel.load_from_folder(os.path.join(training_config.output_dir, f'{trainer.model.model_name}_training_{trainer._training_signature}', 'final_model')).to(device)
+
+    test_data = torch.tensor(test_data).to(device).type(torch.float)
+
+    ### Compute NLL
+    with torch.no_grad():
+        nll = []
+        for i in range(5):
+            nll_i = trained_model.get_nll(test_data, n_samples=500, batch_size=500)
+            logger.info(f"Round {i+1} nll: {nll_i}")
+            nll.append(nll_i)
+    
+    logger.info(
+        f'\nmean_nll: {np.mean(nll)}'
+    )
+    logger.info(
+        f'\std_nll: {np.std(nll)}'
+    )
+
+if __name__ == "__main__":
+
+    main(args)
